+++ PL/TeachRed.cs
@@ -94,8 +94,13 @@ namespace PL
         }
         private void DeleteButt_Click(object sender, EventArgs e)
         {
-            Logic.DeleteButt_Click();
-            Close();
+            DialogResult result = MessageBox.Show("Ви дійсно хочете видалити викладача?", "Видалення",
+                MessageBoxButtons.YesNo, MessageBoxIcon.Question);
+            if (result == DialogResult.Yes)
+            {
+                Logic.DeleteButt_Click();
+                Close();
+            }
         }
         private void TeachRed_FormClosed(object sender, FormClosedEventArgs e)
         {
